Move the `Settings` class to `skills` package
